restructure: move a lot of code to evm \n broke up 2way dependency - added visitor pattern
