--- vms/platformvm/txs/commands/constants.go
@@ -1,7 +1,0 @@
-package commands
-
-import "github.com/ethereum/go-ethereum/common"
-
-var (
-	AdminContractAddr = common.HexToAddress("0x010000000000000000000000000000000000000a")
-)
